+++ src/plugins/config.js
