+++ src/plugins/config.spec.js
