+++ src/plugins/entry.js
