+++ src/plugins/entry.spec.js
