+++ src/plugins/json-data.js
