+++ src/plugins/json-data.spec.js
