+++ src/plugins/route-metadata.js
