+++ src/plugins/route-metadata.spec.js
