+++ src/plugins/template-reference-variable.js
