+++ src/plugins/template-reference-variable.spec.js
